bonita-white-classifier/requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
albumentations>=1.3.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
av>=10.0.0

# Data & ML
scikit-learn>=1.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & Logging
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
pathlib>=1.0.0
python-dotenv>=1.0.0

# Jupyter (para notebooks)
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: para despliegue
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
